refactor: move find shortest pat algorithms to separate class
